Building a Simple RAM Disk Block Driver in Linux
In this post, we'll walk through the a basic RAM disk driver. This driver
allocates a chunk of system memory and presents it to the Linux kernel as
a standard block device. You can then format it with a filesystem (like ext4)
and mount it just like any physical drive.